pAdicSolver.jl Documentation

pAdicSolver.jl Documentation

pAdic Solver is a package for solving 0 dimensional systems of polynomial equations over the p-adics. The method is based on van Baren, Mourrain, Telen's Truncated normal form solver.

Solving

pAdicSolver.macaulay_mat — Method.

macaulay_mat(P::Array{Hecke.Generic.MPoly{T},1},
                  X::Array{Hecke.Generic.MPoly{T},1}, rho, ish) where T <: Hecke.RingElem

Constructs the sparse macaulay matrix defined by the polynomials P and degree bound rho. The argument X is a list of variables of the ambient polynomial ring used to construct the multiplier monomials.

source

pAdicSolver.rauls_change_base_ring — Method.

change_base_ring(p::AbstractAlgebra.MPolyElem{T}, g, R::MPolyRing)
   where T <: RingElement

Return the polynomial in R obtained by applying g to the coefficients of p.

source

pAdicSolver.rel_error — Method.

Gives the list of evaluations of the polynomials in P at the points sol.

source

LinearAlgebra.nullspace — Method.

nullspace(A::Array{padic,2})

Compute the nullspace of an array of padic numbers, viewed as a matrix.

source

pAdicSolver.normalized_simultaneous_eigenvalues — Method.

normalized_simultaneous_eigenvalues(inputM :: Array{Array{T,2},1} where T <: FieldElem, ish::Bool, method)
        --> eigenvalue_matrix :: Hecke.Generic.MatElem{T}

(Internal function) Compute the eigenvalues of a list of commuting matrices, given as an array. In the specific case that the matrices are mult-by-coordinate-variable operators on R/I

INPUTS: M – list of commuting matrices corresponding to mult-by-xi operators Outputs: A matrix whose j-th column are the eigenvalues of the j-th matrix in M

source